fix Conf trait impls
