--- src/configs/types/mod.rs
@@ -1,1 +1,0 @@
-pub mod configs;
