+++ src/gui/types/conf.rs
@@ -7,9 +7,14 @@ use crate::networking::types::capture_context::CaptureSourcePicklist;
 use crate::networking::types::config_device::ConfigDevice;
 use crate::report::types::report_sort_type::ReportSortType;
 use crate::report::types::sort_type::SortType;
+use crate::utils::error_logger::{ErrorLogger, Location};
+use crate::{SNIFFNET_LOWERCASE, location};
+use confy::ConfyError;
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize, Clone, PartialEq, Debug)]
+pub static CONF: std::sync::LazyLock<Conf> = std::sync::LazyLock::new(Conf::load);
+
+#[derive(Serialize, Deserialize, Default, Clone, PartialEq, Debug)]
 pub struct Conf {
     /// Parameters from settings pages
     pub settings: Settings,
@@ -33,4 +38,47 @@ pub struct Conf {
     pub export_pcap: ExportPcap,
     /// Import path for PCAP file
     pub import_pcap_path: String,
+}
+
+impl Conf {
+    const FILE_NAME: &'static str = "conf";
+
+    /// This should only be used directly to load fresh configurations;
+    /// use `CONF` instead to access the initial instance
+    #[cfg(not(test))]
+    pub fn load() -> Self {
+        if let Ok(conf) = confy::load::<Conf>(SNIFFNET_LOWERCASE, Self::FILE_NAME) {
+            conf
+        } else {
+            let _ = confy::store(SNIFFNET_LOWERCASE, Self::FILE_NAME, Conf::default())
+                .log_err(location!());
+            Conf::default()
+        }
+    }
+
+    #[cfg(not(test))]
+    pub fn store(self) -> Result<(), ConfyError> {
+        confy::store(SNIFFNET_LOWERCASE, Self::FILE_NAME, self).log_err(location!())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::Settings;
+    use crate::gui::types::conf::Conf;
+
+    impl Conf {
+        pub fn test_path() -> String {
+            format!("{}/{}.toml", env!("CARGO_MANIFEST_DIR"), Self::FILE_NAME)
+        }
+
+        pub fn load() -> Self {
+            confy::load_path::<Settings>(Settings::test_path())
+                .unwrap_or_else(|_| Settings::default())
+        }
+
+        pub fn store(self) -> Result<(), confy::ConfyError> {
+            confy::store_path(Settings::test_path(), self)
+        }
+    }
 }
